requirements.txt
fastapi~=0.135.3
pydantic~=2.12.5
SQLAlchemy~=2.0.49
python-dotenv~=1.2.2
bcrypt~=5.0.0
passlib~=1.7.4
uvicorn~=0.44.0